graphics/font: use i32 instead of i16 for glyph positions
Using i16 caused a bunch of integer overflows when rendering
slightly longer strings, because coordinates are in font space
(so about 1024 units per em aka character) and i16 can only
handle up to around 30 of those

This doesn't really fix the underlying issue, we will still crash at some point,
however that point now seems to be at around 2 million characters which is much
more reasonable.

+++ graphics/font/src/path.rs
@@ -3,14 +3,14 @@ use math::Vec2D;
 
 #[derive(Clone, Copy, Debug)]
 pub enum Operation {
-    LineTo(Vec2D<i16>),
-    QuadBezTo(Vec2D<i16>, Vec2D<i16>),
-    MoveTo(Vec2D<i16>),
+    LineTo(Vec2D<i32>),
+    QuadBezTo(Vec2D<i32>, Vec2D<i32>),
+    MoveTo(Vec2D<i32>),
 }
 
 pub struct PathReader<I: Iterator<Item = GlyphPoint>> {
     inner: I,
-    last_on_curve_point: Option<Vec2D<i16>>,
+    last_on_curve_point: Option<Vec2D<i32>>,
     previous_point: Option<GlyphPoint>,
     first_point_of_contour: Option<GlyphPoint>,
     state: PathReaderState,
